+++ src/node/Bitmap.ts
@@ -938,7 +938,7 @@ class Bitmap extends Node {
     this.loadAndRefresh();
   }
 
-  override get bbox(): Float64Array {
+  override get bbox(): Float32Array {
     let res = this._bbox;
     if (!res) {
       const rect = this._rect || this.rect;
